web/requirements/base.txt
Django==1.10.6
psycopg2
gunicorn==19.7.1
unipath>=1.1
django-allauth
django-import-export
django-adaptors
django-csvimport
